+++ src/components/DisplayText/DisplayText.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {classNames, variationName} from '../../utilities/css';
-import {HeadingTagName} from '../../types';
+import type {HeadingTagName} from '../../types';
 
 import styles from './DisplayText.scss';