+++ src/components/Filters/components/ConnectedFilterControl/ConnectedFilterControl.tsx
@@ -3,7 +3,7 @@ import debounce from 'lodash/debounce';
 
 import {classNames} from '../../../../utilities/css';
 import {FeaturesContext} from '../../../../utilities/features';
-import {DisableableAction} from '../../../../types';
+import type {DisableableAction} from '../../../../types';
 import {Popover} from '../../../Popover';
 import {Button} from '../../../Button';
 import {EventListener} from '../../../EventListener';