+++ src/components/ResourceList/components/BulkActions/BulkActions.tsx
@@ -4,7 +4,11 @@ import {durationBase} from '@shopify/polaris-tokens';
 import {CSSTransition, Transition} from '@material-ui/react-transition-group';
 
 import {classNames} from '../../../../utilities/css';
-import {DisableableAction, Action, ActionListSection} from '../../../../types';
+import type {
+  DisableableAction,
+  Action,
+  ActionListSection,
+} from '../../../../types';
 import {ActionList} from '../../../ActionList';
 import {Popover} from '../../../Popover';
 import {Button} from '../../../Button';