+++ src/utilities/scroll-lock-manager/context.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {ScrollLockManager} from './scroll-lock-manager';
+import type {ScrollLockManager} from './scroll-lock-manager';
 
 export const ScrollLockManagerContext = React.createContext<
   ScrollLockManager | undefined